Minor changes and tweaks after merge with main
