+++ src/Server.UI/Components/Shared/Layout/SearchDialog.razor
@@ -85,7 +85,7 @@
                 pages.Add("Reassign", "/pages/participants/reassign");
                 pages.Add("Transfers", "/pages/participants/transfers");
                 pages.Add("Moved Participants", "/pages/participants/movedparticipants");
-                pages.Add("Active PRI's", "/pages/participants/activePris");
+                pages.Add("Active PRI's", "/pages/participants/pre-release-inventory");
             }
 
             string[] allowed = [RoleNames.QAFinance, RoleNames.SMT, RoleNames.SystemSupport];
